moved classes of arguments, dir listing and jobs to their corresponding directories
